вынес бизнес-логику из моделей в репозитории, добавил счетчик просмотров, изменил лайки на оценку от 1 до 5, добавил таблицу с настройками сайта, рефакторинг

+++ console/migrations/m210526_082440_create_gallery_table.php
@@ -21,7 +21,7 @@ class m210526_082440_create_gallery_table extends Migration
                     'name' => $this->string()->notNull(),
                     'category_id' => $this->integer()->notNull(),
                     'created_at' => $this->dateTime()->notNull(),
-                    'rating' => $this->integer()->notNull()->defaultValue(0),
+                    'rating' => $this->double()->notNull()->defaultValue(0),
                     'moderation' => $this->integer()->notNull()->defaultValue(0),
                 ],
             );
